requirements.txt
<<<<<<< HEAD
requests
python-dotenv
=======
requests
python-dotenv
>>>>>>> 8ccd15582954fdb03959ebeffec6ff339ed5b2c1
